Extract name-mangling logic to a new header
