+++ packager/packager.cc
@@ -13,6 +13,7 @@
 #include "core/AutocorrectSuggestion.h"
 #include "core/Unfreeze.h"
 #include "core/errors/packager.h"
+#include "core/packages/MangledName.h"
 #include "core/packages/PackageInfo.h"
 #include <algorithm>
 #include <cctype>
@@ -67,20 +68,6 @@ struct FullyQualifiedName {
     }
 };
 
-class NameFormatter final {
-    const core::GlobalState &gs;
-
-public:
-    NameFormatter(const core::GlobalState &gs) : gs(gs) {}
-
-    void operator()(std::string *out, core::NameRef name) const {
-        out->append(name.shortName(gs));
-    }
-    void operator()(std::string *out, pair<core::NameRef, core::LocOffsets> p) const {
-        out->append(p.first.shortName(gs));
-    }
-};
-
 struct PackageName {
     core::LocOffsets loc;
     core::NameRef mangledName = core::NameRef::noName();
@@ -89,7 +76,7 @@ struct PackageName {
 
     // Pretty print the package's (user-observable) name (e.g. Foo::Bar)
     string toString(const core::GlobalState &gs) const {
-        return absl::StrJoin(fullName.parts, "::", NameFormatter(gs));
+        return absl::StrJoin(fullName.parts, "::", core::packages::NameFormatter(gs));
     }
 
     bool operator==(const PackageName &rhs) const {
@@ -438,11 +425,7 @@ PackageName getPackageName(core::MutableContext ctx, ast::UnresolvedConstantLit 
     pName.fullTestPkgName = pName.fullName.withPrefix(TEST_NAME);
 
     // Foo::Bar => Foo_Bar_Package
-    auto mangledName = absl::StrCat(absl::StrJoin(pName.fullName.parts, "_", NameFormatter(ctx)), core::PACKAGE_SUFFIX);
-
-    auto utf8Name = ctx.state.enterNameUTF8(mangledName);
-    auto packagerName = ctx.state.freshNameUnique(core::UniqueNameKind::Packager, utf8Name, 1);
-    pName.mangledName = ctx.state.enterNameConstant(packagerName);
+    pName.mangledName = core::packages::MangledName::mangledNameFromParts(ctx, pName.fullName.parts);
 
     return pName;
 }
@@ -917,7 +900,8 @@ private:
         auto reqMangledName = namespaces.packageForNamespace();
         if (reqMangledName.exists()) {
             auto &reqPkg = gs.packageDB().getPackageInfo(reqMangledName);
-            auto givenNamespace = absl::StrJoin(namespaces.currentConstantName(), "::", NameFormatter(gs));
+            auto givenNamespace =
+                absl::StrJoin(namespaces.currentConstantName(), "::", core::packages::NameFormatter(gs));
             e.addErrorLine(reqPkg.declLoc(), "Must belong to this package, given constant name `{}`", givenNamespace);
         }
     }
@@ -1618,7 +1602,7 @@ public:
     ImportFormatter(const core::GlobalState &gs) : gs(gs) {}
 
     void operator()(std::string *out, const vector<core::NameRef> &name) const {
-        fmt::format_to(back_inserter(*out), "\"{}\"", absl::StrJoin(name, "::", NameFormatter(gs)));
+        fmt::format_to(back_inserter(*out), "\"{}\"", absl::StrJoin(name, "::", core::packages::NameFormatter(gs)));
     }
 };
 
@@ -1652,7 +1636,8 @@ public:
         const auto &pkg = gs.packageDB().getPackageInfo(mangledName);
         out->append("{{");
         out->append("\"name\":");
-        fmt::format_to(back_inserter(*out), "\"{}\",", absl::StrJoin(pkg.fullName(), "::", NameFormatter(gs)));
+        fmt::format_to(back_inserter(*out), "\"{}\",",
+                       absl::StrJoin(pkg.fullName(), "::", core::packages::NameFormatter(gs)));
         out->append("\"imports\":[");
         fmt::format_to(back_inserter(*out), absl::StrJoin(pkg.imports(), ",", ImportFormatter(gs)));
         out->append("],\"testImports\":[");
